requirements.txt
requests==2.25.1
psutil==5.7.2
wget==3.2
appdirs==1.4.4
python_dateutil==2.8.2
oschmod
pyyaml
py-dactyl